multimodal_server/requirements.txt
fastapi==0.110.0
uvicorn==0.28.1
httpx>=0.24.0
requests>=2.31.0
websocket-client>=1.6.0
loguru==0.7.2

Pillow==10.2.0
imagehash==4.3.1

torch>=2.0.0
transformers==4.38.2
accelerate==0.28.0
diffusers[torch]
safetensors==0.4.2
einops==0.7.0
kornia>=0.7.1
python-dotenv

numpy>=1.21.0
scipy==1.10.1
pyyaml==6.0.1
ftfy==6.2.0
regex==2023.12.25
tqdm==4.66.2

aiohttp==3.9.3
psutil==5.9.8

gdown==5.1.0
torchsde==0.2.6
insightface==0.7.3
onnxruntime==1.17.1
protobuf==5.27.0
huggingface_hub[hf_transfer]==0.23.2
open-clip-torch>=2.24.0
pytorch-lightning>=2.2.1
omegaconf==2.3.0

# CLIP
git+https://github.com/openai/CLIP.git
